isolate Advanced Settings to its own component
