+++ src/containers/AccessPointDetails/index.module.scss
@@ -34,87 +34,87 @@
   :global(.ant-form-item) {
     flex: 1;
   }
+}
 
-  .InlineBlockDiv {
-    display: inline-block;
-    vertical-align: middle;
-    :global(.ant-avatar) {
-      margin-right: 20px;
-    }
-  }
-
-  .WifiIcon {
-    background-color: transparent;
-    color: #35a649;
-    border: none;
-    font-size: 40px;
-    vertical-align: middle;
+.InlineBlockDiv {
+  display: inline-block;
+  vertical-align: middle;
+  :global(.ant-avatar) {
     margin-right: 20px;
   }
+}
 
-  .backButton {
-    margin-top: 5px;
-  }
+.WifiIcon {
+  background-color: transparent;
+  color: #35a649;
+  border: none;
+  font-size: 40px;
+  vertical-align: middle;
+  margin-right: 20px;
+}
 
-  .InlineDiv {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-  }
+.backButton {
+  margin-top: 5px;
+}
 
-  .InlineDiv > * {
-    margin: 0 10px 0 0;
-  }
+.InlineDiv {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+}
 
-  .InlineBetweenDiv {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
+.InlineDiv > * {
+  margin: 0 10px 0 0;
+}
 
-  .saveButton {
-    margin-top: 10px;
-    min-width: 139px;
-    max-width: 100%;
-  }
+.InlineBetweenDiv {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
 
-  .InlineEndDiv {
-    display: flex;
-    justify-content: flex-end;
-  }
+.saveButton {
+  margin-top: 10px;
+  min-width: 139px;
+  max-width: 100%;
+}
 
-  p {
-    font-weight: bold !important;
-    font-size: 16px;
-  }
-  .spanStyle {
-    flex: 1;
-    margin-left: 10px;
-  }
+.InlineEndDiv {
+  display: flex;
+  justify-content: flex-end;
+}
 
-  .UpgradeState {
-    margin-left: 20px;
-  }
+p {
+  font-weight: bold !important;
+  font-size: 16px;
+}
+.spanStyle {
+  flex: 1;
+  margin-left: 10px;
+}
 
-  .HeaderDiv {
-    display: flex;
-  }
+.UpgradeState {
+  margin-left: 20px;
+}
 
-  .troubleshootBtnsDiv {
-    & > * {
-      margin-right: 10px;
-    }
-  }
+.HeaderDiv {
+  display: flex;
+}
 
-  .tabContentContainer {
-    margin-top: 10px;
+.troubleshootBtnsDiv {
+  & > * {
+    margin-right: 10px;
   }
+}
 
-  .Row {
-    cursor: pointer;
-  }
+.tabContentContainer {
+  margin-top: 10px;
+}
 
-  a {
-    color: white;
-  }
+.Row {
+  cursor: pointer;
+}
+
+a {
+  color: white;
 }
